# backend/src/main/resources/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PlayerSignal local alpha",
    "version": "0.6.0",
    "description": "Accounts are opt-in via AUTH_ENABLED. When enabled, game resources require a session and every mutation requires X-CSRF-TOKEN from /api/auth/session. Cross-workspace resources return 404."
  },
  "paths": {
    "/api/games": {
      "get": {
        "summary": "List connected games",
        "responses": {
          "200": {
            "description": "Games",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Game"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Resolve App ID or HTTPS Steam store URL and upsert game",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "steamApp"
                ],
                "properties": {
                  "steamApp": {
                    "type": "string",
                    "maxLength": 2048
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Import already running or capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Steam unavailable or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Import could not start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required when accounts are enabled.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/games/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Get connected game",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Import already running or capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Steam unavailable or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Import could not start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/games/{id}/sync": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "summary": "Start bounded background import; resume previous incomplete run",
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Import already running or capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Steam unavailable or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Import could not start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required when accounts are enabled.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/games/{id}/sync/latest": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Latest import audit",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "204": {
            "description": "No import has run yet"
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Import already running or capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Steam unavailable or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Import could not start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/games/{id}/reviews": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Browse persisted source reviews",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000,
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^[a-z]{2,32}$"
            }
          },
          {
            "name": "votedUp",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 256
            },
            "description": "Literal case-insensitive substring of original review text"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Import already running or capacity exhausted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Steam unavailable or invalid response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Import could not start",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/games/{id}/analysis": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Current analysis configuration, counts and latest run",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatus"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Start a bounded analysis batch",
        "parameters": [
          {
            "name": "retryFailed",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Required when accounts are enabled.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRun"
                }
              }
            }
          },
          "409": {
            "description": "Analysis worker or game busy"
          },
          "503": {
            "description": "Provider unconfigured or unavailable"
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/games/{id}/reviews/{reviewId}/analyses": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "reviewId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "All analysis versions with exact input snapshots",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisHistory"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Review not found in this game"
          },
          "401": {
            "description": "Authentication required when accounts are enabled"
          }
        },
        "security": [
          {},
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "required": [
          "id",
          "steamAppId",
          "name",
          "headerImageUrl",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "steamAppId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "headerImageUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Run": {
        "type": "object",
        "required": [
          "id",
          "gameId",
          "status",
          "startedAt",
          "finishedAt",
          "fetched",
          "inserted",
          "updated",
          "nextCursor",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "FAILED"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fetched": {
            "type": "integer",
            "minimum": 0
          },
          "inserted": {
            "type": "integer",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Review": {
        "type": "object",
        "required": [
          "id",
          "steamRecommendationId",
          "language",
          "reviewText",
          "votedUp",
          "votesUp",
          "playtimeMinutes",
          "createdAtSteam",
          "updatedAtSteam"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "steamRecommendationId": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "reviewText": {
            "type": "string"
          },
          "votedUp": {
            "type": "boolean"
          },
          "votesUp": {
            "type": "integer",
            "minimum": 0
          },
          "playtimeMinutes": {
            "type": "integer",
            "minimum": 0
          },
          "createdAtSteam": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtSteam": {
            "type": "string",
            "format": "date-time"
          },
          "analysis": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Analysis"
              }
            ],
            "nullable": true
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "ReviewPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Classification": {
        "type": "object",
        "properties": {
          "sentiment": {
            "type": "string",
            "enum": [
              "VERY_NEGATIVE",
              "NEGATIVE",
              "MIXED",
              "POSITIVE",
              "VERY_POSITIVE"
            ]
          },
          "primaryCategory": {
            "type": "string",
            "enum": [
              "BUG",
              "PERFORMANCE",
              "GAMEPLAY",
              "UI_UX",
              "MULTIPLAYER",
              "BALANCE",
              "CONTENT",
              "CONTROLS",
              "AUDIO",
              "LOCALIZATION",
              "POSITIVE",
              "OTHER"
            ]
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ]
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "normalizedIssue": {
            "type": "string",
            "maxLength": 240
          },
          "isActionable": {
            "type": "boolean"
          },
          "isLikelyBug": {
            "type": "boolean"
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 500
            },
            "maxItems": 5
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 40
            },
            "maxItems": 8
          }
        },
        "required": [
          "sentiment",
          "primaryCategory",
          "severity",
          "confidence",
          "normalizedIssue",
          "isActionable",
          "isLikelyBug",
          "evidence",
          "tags"
        ],
        "additionalProperties": false
      },
      "Analysis": {
        "type": "object",
        "required": [
          "id",
          "provider",
          "model",
          "promptVersion",
          "updatedAt",
          "responseModel",
          "error",
          "skipReason",
          "cachedFrom",
          "status",
          "attempts",
          "result"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "responseModel": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "skipReason": {
            "type": "string",
            "nullable": true
          },
          "cachedFrom": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SUCCEEDED",
              "FAILED",
              "SKIPPED"
            ]
          },
          "attempts": {
            "type": "integer",
            "minimum": 0
          },
          "result": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Classification"
              }
            ],
            "nullable": true
          }
        }
      },
      "AnalysisRun": {
        "type": "object",
        "required": [
          "id",
          "gameId",
          "provider",
          "model",
          "promptVersion",
          "startedAt",
          "status",
          "processed",
          "succeeded",
          "skipped",
          "failed",
          "cached",
          "finishedAt",
          "error"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "gameId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "startedAt": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "COMPLETED",
              "PARTIAL",
              "COMPLETED_WITH_ERRORS",
              "FAILED"
            ]
          },
          "processed": {
            "type": "integer",
            "minimum": 0
          },
          "succeeded": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "integer",
            "minimum": 0
          },
          "failed": {
            "type": "integer",
            "minimum": 0
          },
          "cached": {
            "type": "integer",
            "minimum": 0
          },
          "finishedAt": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AnalysisStatus": {
        "type": "object",
        "required": [
          "available",
          "provider",
          "model",
          "promptVersion",
          "maxReviewsPerRun",
          "counts",
          "latestRun"
        ],
        "properties": {
          "available": {
            "type": "boolean"
          },
          "provider": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "maxReviewsPerRun": {
            "type": "integer",
            "minimum": 0
          },
          "counts": {
            "type": "object",
            "required": [
              "total",
              "pending",
              "running",
              "succeeded",
              "skipped",
              "failed"
            ],
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0
              },
              "pending": {
                "type": "integer",
                "minimum": 0
              },
              "running": {
                "type": "integer",
                "minimum": 0
              },
              "succeeded": {
                "type": "integer",
                "minimum": 0
              },
              "skipped": {
                "type": "integer",
                "minimum": 0
              },
              "failed": {
                "type": "integer",
                "minimum": 0
              }
            }
          },
          "latestRun": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AnalysisRun"
              }
            ],
            "nullable": true
          }
        }
      },
      "AnalysisHistory": {
        "type": "object",
        "required": [
          "analysis",
          "inputHash",
          "sourceText",
          "sourceLanguage"
        ],
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/Analysis"
          },
          "inputHash": {
            "type": "string"
          },
          "sourceText": {
            "type": "string"
          },
          "sourceLanguage": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "PLAYERSIGNAL_SESSION"
      }
    }
  }
}